Create chat component. Implement websockets for timely API calls.
Websockets are implemented as a basic overlay to the existing
routing. Rather than refactoring, I opted to use websockets
to notify the server that the other connected clients needed
to update. This then notifies the clients that run the normal
http requests. Not the most streamlined method, but it works.

Chat is implimented in roughly the same fashion.